+++ src/constructor.js
@@ -20,6 +20,32 @@ function Tenso () {
 Tenso.prototype.constructor = Tenso;
 
 /**
+ * Sends an Error to the Client
+ *
+ * @method redirect
+ * @memberOf Tenso
+ * @param  {Object} req Client request
+ * @param  {Object} res Client response
+ * @param  {Mixed}  uri Target URI
+ */
+Tenso.prototype.error = function ( req, res, status, arg ) {
+	this.server.error( req, res, status, arg );
+};
+
+/**
+ * Redirects the Client
+ *
+ * @method redirect
+ * @memberOf Tenso
+ * @param  {Object} req Client request
+ * @param  {Object} res Client response
+ * @param  {Mixed}  uri Target URI
+ */
+Tenso.prototype.redirect = function ( req, res, uri ) {
+	this.server.respond( req, res, this.server.messages.NO_CONTENT, this.server.codes.FOUND, {location: uri} );
+};
+
+/**
  * Sends a response to the Client
  *
  * @method respond
@@ -34,5 +60,7 @@ Tenso.prototype.constructor = Tenso;
 Tenso.prototype.respond = function ( req, res, arg, status, headers ) {
 	var ref = [headers || {}];
 
-	this.server.respond( req, res, hypermedia( this.server, req, response( arg, status ), ref[0] ), status, ref[0] );
+	if ( !res._header ) {
+		this.server.respond( req, res, hypermedia( this.server, req, response( arg, status ), ref[0] ), status, ref[0] );
+	}
 };